Handle notebook selection changes via keyboard events
QTreeView (and QAbstractItemView) don't export a signal for selection
changed events.  This makes it difficult to capture keyboard events and
handle selection change notifications appropriately.

This change creates a new subclass of QTreeView, EverpadTreeView, that
overrides the selectionChanged method and emits a signal, so we can
properly update the notes list, no matter how the selection was changed.